Fixed typo
There was a typo in the require statement in artoo-arduino. This wasn't
being picked up by the tests because it wasn't being required anywhere.
Requiring caused some additional issues with other required files being
picked up so these have been updated.

In addition it was requiring the littlewire adapter which was extracted
into its own project in 2a3fa12.